Split Kernel into smaller files
