+++ meerk40t/core/bindalias.py
@@ -1,4 +1,5 @@
-from ..kernel import CommandMatchRejected, Service
+from meerk40t.kernel.service import Service
+from meerk40t.core.exceptions import CommandMatchRejected
 
 # The following dicts consist of a tuple of values, the first of
 # which is the current default, with any prior defaults following.
