+++ meerk40t/core/exceptions.py
@@ -14,3 +14,13 @@ class Mk40tImportAbort(ImportError, Mk40tError):
     except ImportError as e:
         raise Mk40tImportAbort("wx") from e
     """
+
+class CommandMatchRejected(Mk40tError):
+    """
+    Exception to be raised by a registered console command if the match to the command was erroneous
+    """
+
+class MalformedCommandRegistration(Mk40tError):
+    """
+    Exception raised by the Kernel if the registration of the console command is malformed.
+    """
