+++ meerk40t/grbl/device.py
@@ -27,7 +27,7 @@ from ..device.basedevice import (
     PLOT_SETTING,
     PLOT_START,
 )
-from ..kernel import Service
+from meerk40t.kernel.service import Service
 
 STATE_ABORT = -1
 STATE_DEFAULT = 0
