+++ meerk40t/grbl/gui/grblserialcontroller.py
@@ -10,7 +10,7 @@ _ = wx.GetTranslation
 
 from meerk40t.gui.icons import icons8_connected_50, icons8_disconnected_50
 from meerk40t.gui.mwindow import MWindow
-from meerk40t.kernel import signal_listener
+from meerk40t.kernel.kernel import signal_listener
 
 
 class SerialControllerPanel(wx.Panel):
