+++ meerk40t/gui/operationpropertymain.py
@@ -1,6 +1,6 @@
 import wx
 
-from ..kernel import signal_listener
+from meerk40t.kernel.kernel import signal_listener
 from ..svgelements import Color
 from .laserrender import swizzlecolor
