+++ meerk40t/gui/wxmribbon.py
@@ -4,7 +4,8 @@ import wx
 import wx.ribbon as RB
 from wx import aui
 
-from ..kernel import Job, lookup_listener
+from meerk40t.kernel.kernel import lookup_listener
+from meerk40t.kernel.jobs import Job
 from .icons import icons8_connected_50, icons8_opened_folder_50
 from .mwindow import MWindow
