+++ meerk40t/gui/wxmscene.py
@@ -27,7 +27,7 @@ from meerk40t.gui.scene.toolwidgets import (
     VectorTool,
 )
 from meerk40t.gui.wxutils import get_key_name
-from meerk40t.kernel import signal_listener
+from meerk40t.kernel.kernel import signal_listener
 from meerk40t.svgelements import Angle, Length
 
 _ = wx.GetTranslation
