+++ meerk40t/kernel/states.py
@@ -1,0 +1,10 @@
+STATE_UNKNOWN = -1
+STATE_INITIALIZE = 0
+STATE_IDLE = 1
+STATE_ACTIVE = 2
+STATE_BUSY = 3
+STATE_PAUSE = 4
+STATE_END = 5
+STATE_SUSPEND = 6
+STATE_WAIT = 7
+STATE_TERMINATE = 10
