+++ meerk40t/lihuiyu/gui/lhystudiosdrivergui.py
@@ -4,7 +4,7 @@ import wx
 
 from meerk40t.gui.icons import icons8_administrative_tools_50
 from meerk40t.gui.mwindow import MWindow
-from meerk40t.kernel import signal_listener
+from meerk40t.kernel.kernel import signal_listener
 
 _ = wx.GetTranslation
