+++ meerk40t/lihuiyu/gui/tcpcontroller.py
@@ -2,7 +2,7 @@ import wx
 
 from meerk40t.gui.icons import icons8_connected_50, icons8_disconnected_50
 from meerk40t.gui.mwindow import MWindow
-from meerk40t.kernel import signal_listener
+from meerk40t.kernel.kernel import signal_listener
 
 _ = wx.GetTranslation
